+++ crates/abe/src/eval.rs
@@ -3,9 +3,6 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
-/**
-This stuff is a mess.
-**/
 use std::fmt::Write;
 use std::ops::{Try, ControlFlow};