+++ emulate/session.tmux.sh
@@ -14,6 +14,6 @@ echo $name $(lk key --no-enckey --no-check) | tee ./emulate_name_key
 tmux -S tmux-socket new-session -s $name -n $name \; \
      send-keys " lk " \; \
      split-window -v  \; \
-     send-keys "lk --private watch :: --bare --new | lk printf" C-m \; \
+     send-keys "lk --private watch :: --bare --new | lk pktf" C-m \; \
      split-window -v \; \
      send-keys "$cmd" C-m