+++ examples/imageboard/imageboard.view.sh
@@ -8,14 +8,14 @@ START_STAMP=${2:-"0"} # If no stamp is given we begin at 0, i.e. unix epoch in m
 
 # We select everything with a create field greater or equal to $START_STAMP
 lk watch --index "imageboard:$LK_GROUP:/$BOARD" -- "create:>=:[u64:$START_STAMP]" \
-    | lk printf "[/links:[tag:str] [ptr:str]]" \
+    | lk pktf "[/links:[tag:str] [ptr:str]]" \
     | while read REF; do
         X=${REF:0:8}
         Y=${REF:8:8}
         IMG_HASH=${REF: -43}
         echo "Placing $IMG_HASH at $X , $Y"
         lk watch-hash $IMG_HASH --ttl 5s \
-            | lk printf "[data]" --delimiter "" \
+            | lk pktf "[data]" --delimiter "" \
             | magick composite -geometry +$X+$Y - PNG32:$BOARD.png PNG32:$BOARD.png
     done
 echo "$BOARD: $START_STAMP"